+++ safeway_coupons/errors.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 from typing import List
 
-from .account import Account
+from .accounts import Account
 from .models import Offer
